Rename `'paq-nvim'` module to `'paq'`
